refactor: simplify logic for extracting quotedTweets
I create a simplier method for extracting quotedTweets.

Since Parquet is going to be used, I only need to have all the tweets in one parquet file, to then just read it and select the columns that I need
